staff-service/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
PyJWT>=2.8
